+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/UIDefaults.java
@@ -9,4 +9,7 @@ public interface UIDefaults {
 
   int THUMBNAIL_SIZE = 240;
   int SCROLL_OFFSET = 120;
+
+
+  int MAX_REFRESH_COUNT = 3;
 }