+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/frontend/popper/PinUPConnector.java
@@ -1215,6 +1215,8 @@ public class PinUPConnector implements FrontendConnector, InitializingBean {
     e.setDirRoms(rs.getString("DirRoms"));
     e.setDescription(rs.getString("Description"));
     e.setEmuLaunchDir(rs.getString("EmuLaunchDir"));
+    e.setLaunchScript(rs.getString("LaunchScript"));
+    e.setExitScript(rs.getString("PostScript"));
     e.setVisible(rs.getInt("Visible") == 1);
 
     // specific initialization