+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TableDialogs.java
@@ -311,6 +311,14 @@ public class TableDialogs {
     return true;
   }
 
+  public static boolean openHighscoresResetDialog(List<GameRepresentation> games) {
+    Stage stage = Dialogs.createStudioDialogStage(HighscoreResetController.class, "dialog-highscore-reset.fxml", "Reset Highscores");
+    HighscoreResetController controller = (HighscoreResetController) stage.getUserData();
+    controller.setGames(games);
+    stage.showAndWait();
+    return true;
+  }
+
 
   public static boolean openNotesDialog(GameRepresentation game) {
     Stage stage = Dialogs.createStudioDialogStage(TableNotesController.class, "dialog-table-notes.fxml", "Comments");